Did the UserProfileActivity that shows a user's public profile. Renamed the old UserProfileActivity to LocalUsersProfileActivity. Renamed the old SimpleUser to LocalUser
It doesn't have hobbies in the profile or rankings yet but I provided an example of how the users hobbies and ranks could show up on his profile.

Also, I couldn't manage to connect to the db through the sockets and retrieve a user by using UUID, mainly (I think) because I don't understand how requests to the server are made however this should be easily fixed by someone explaining to me how the server works.