The user can now select the place using the map and accept. After this, the address of the place will be displayed.
Important: The "Event" Json will no longer contain the address in location but rather the place ID. This should be ok since we never specified in the rfc what it contained. I don't know how this will work with gogo deals however. It should be beneficial for them I guess-